cloudfunctions/src/bigquery/updatetable/requirements.txt
# Function dependencies, for example:
# package>=version
sodapy
google-cloud-storage
google-cloud-bigquery